Refining the move from old Dashboard to new
refs: https://github.com/TryGhost/Team/issues/1631
