+++ app/components/dashboard/charts/engagement.hbs
@@ -1,24 +1,24 @@
-<section class="gh-dashboard5-section gh-dashboard5-engagement">
-    <article {{did-insert this.loadCharts}} class="gh-dashboard5-box">
+<section class="gh-dashboard-section gh-dashboard-engagement">
+    <article {{did-insert this.loadCharts}} class="gh-dashboard-box">
         <Dashboard::Parts::Metric
             @label="Engagement" /> 
 
-        <div class="gh-dashboard5-columns">
-            <div class="gh-dashboard5-column gh-dashboard5-engagement-30days">
+        <div class="gh-dashboard-columns">
+            <div class="gh-dashboard-column gh-dashboard-engagement-30days">
                 <Dashboard::Parts::Metric
                     @label="Engaged in the last 30 days"
                     @value={{this.data30Days}}
                     @secondary={{true}}
                     />
             </div>
-            <div class="gh-dashboard5-column gh-dashboard5-engagement-7days">
+            <div class="gh-dashboard-column gh-dashboard-engagement-7days">
                 <Dashboard::Parts::Metric
                     @label="Engaged in the last 7 days"
                     @value={{this.data7Days}}
                     @secondary={{true}}
                     />
             </div>
-            <div class="gh-dashboard5-column gh-dashboard5-engagement-subscribers">
+            <div class="gh-dashboard-column gh-dashboard-engagement-subscribers">
                 <Dashboard::Parts::Metric
                     @label="Newsletter subscribers"
                     @value={{this.dataSubscribers}}
@@ -29,7 +29,7 @@
     </article>
 
     {{#if this.hasPaidTiers}}
-        <div class="gh-dashboard5-select">
+        <div class="gh-dashboard-select">
             <PowerSelect
                 @selected={{this.selectedStatusOption}}
                 @options={{this.statusOptions}}
