+++ app/components/dashboard/charts/overview.hbs
@@ -1,7 +1,7 @@
-<section class="gh-dashboard5-section gh-dashboard5-overview {{unless this.isTotalMembersMoreThanZero 'is-hidden'}}">
-    <article {{did-insert this.loadCharts}} class="gh-dashboard5-box is-secondary">
-        <div class="gh-dashboard5-columns">
-            <div class="gh-dashboard5-column">
+<section class="gh-dashboard-section gh-dashboard-overview {{unless this.isTotalMembersMoreThanZero 'is-hidden'}}">
+    <article {{did-insert this.loadCharts}} class="gh-dashboard-box is-secondary">
+        <div class="gh-dashboard-columns">
+            <div class="gh-dashboard-column">
                 <Dashboard::Parts::Metric
                     @label={{gh-pluralize this.totalMembers "Total member" without-count=true}}
                     @value={{this.totalMembersFormatted}}
@@ -9,7 +9,7 @@
                     @percentage={{this.totalMembersTrend}}
                     @large={{true}} /> 
             </div>
-            <div class="gh-dashboard5-column">
+            <div class="gh-dashboard-column">
                 <Dashboard::Parts::Metric
                     @label={{gh-pluralize this.paidMembers "Paid member" without-count=true}}
                     @value={{this.paidMembersFormatted}}
@@ -17,7 +17,7 @@
                     @percentage={{this.paidMembersTrend}}
                     @large={{true}} /> 
             </div>
-            <div class="gh-dashboard5-column">
+            <div class="gh-dashboard-column">
                 <Dashboard::Parts::Metric
                     @label={{gh-pluralize this.freeMembers "Free member" without-count=true}}
                     @value={{this.freeMembersFormatted}}
