+++ app/components/dashboard/charts/paid-breakdown.hbs
@@ -1,22 +1,22 @@
-<div class="gh-dashboard5-minichart gh-dashboard5-breakdown">
-    <div class="gh-dashboard5-content">
-        <div class="gh-dashboard5-data">
+<div class="gh-dashboard-minichart gh-dashboard-breakdown">
+    <div class="gh-dashboard-content">
+        <div class="gh-dashboard-data">
             <Dashboard::Parts::Metric
                 @label={{this.chartTitle}} />
 
-            <div class="gh-dashboard5-legend">
-                <div class="gh-dashboard5-legend-item">New</div>
-                <div class="gh-dashboard5-legend-item">Canceled</div>
+            <div class="gh-dashboard-legend">
+                <div class="gh-dashboard-legend-item">New</div>
+                <div class="gh-dashboard-legend-item">Canceled</div>
             </div>
         </div>
-        <div class="gh-dashboard5-chart" {{did-insert this.loadCharts}}>
+        <div class="gh-dashboard-chart" {{did-insert this.loadCharts}}>
             {{#if this.loading}}
-                <div class="gh-dashboard5-chart-loading">
+                <div class="gh-dashboard-chart-loading">
                     <div class="gh-loading-spinner"></div>
                 </div>
             {{else}}
-                <div class="gh-dashboard5-chart-container">
-                    <div class="gh-dashboard5-chart-box">
+                <div class="gh-dashboard-chart-container">
+                    <div class="gh-dashboard-chart-box">
                         <EmberChart
                             @type={{this.chartType}}
                             @data={{this.chartData}}
@@ -24,16 +24,16 @@
                             @height={{110}} />
                     </div>
 
-                    <div id="gh-dashboard5-breakdown-tooltip" class="gh-dashboard5-tooltip">
-                        <div class="gh-dashboard5-tooltip-label">
+                    <div id="gh-dashboard-breakdown-tooltip" class="gh-dashboard-tooltip">
+                        <div class="gh-dashboard-tooltip-label">
                             -
                         </div>
-                        <div class="gh-dashboard5-tooltip-value">
-                            <div class="gh-dashboard5-tooltip-value-1"><span class="indicator solid"></span><span class="value"></span></div>
-                            <div class="gh-dashboard5-tooltip-value-1"><span class="metric">New</span></div>
-                            <div class="gh-dashboard5-tooltip-value-2"><span class="indicator solid"></span><span class="value"></span></div>
-                            <div class="gh-dashboard5-tooltip-value-2"><span class="metric">Canceled</span></div>
-                            <div class="gh-dashboard5-tooltip-value-3"></div>
+                        <div class="gh-dashboard-tooltip-value">
+                            <div class="gh-dashboard-tooltip-value-1"><span class="indicator solid"></span><span class="value"></span></div>
+                            <div class="gh-dashboard-tooltip-value-1"><span class="metric">New</span></div>
+                            <div class="gh-dashboard-tooltip-value-2"><span class="indicator solid"></span><span class="value"></span></div>
+                            <div class="gh-dashboard-tooltip-value-2"><span class="metric">Canceled</span></div>
+                            <div class="gh-dashboard-tooltip-value-3"></div>
                         </div>
                     </div> 
                 </div>
