+++ app/components/dashboard/charts/paid-mix.hbs
@@ -1,24 +1,24 @@
-<div class="gh-dashboard5-minichart gh-dashboard5-mix {{if this.isChartTiers 'is-tiers'}}">
-    <div class="gh-dashboard5-content">
-        <div class="gh-dashboard5-data">
+<div class="gh-dashboard-minichart gh-dashboard-mix {{if this.isChartTiers 'is-tiers'}}">
+    <div class="gh-dashboard-content">
+        <div class="gh-dashboard-data">
             <Dashboard::Parts::Metric
                 @label="Paid mix" /> 
 
             {{#if this.isChartCadence}}
-                <div class="gh-dashboard5-legend">
-                    <div class="gh-dashboard5-legend-item">Monthly</div>
-                    <div class="gh-dashboard5-legend-item">Annual</div>
+                <div class="gh-dashboard-legend">
+                    <div class="gh-dashboard-legend-item">Monthly</div>
+                    <div class="gh-dashboard-legend-item">Annual</div>
                 </div>
             {{/if}}
         </div>
-        <div class="gh-dashboard5-chart {{if this.isChartCadence "narrow"}}" {{did-insert this.loadCharts}}>
+        <div class="gh-dashboard-chart {{if this.isChartCadence "narrow"}}" {{did-insert this.loadCharts}}>
             {{#if this.loading}}
-                <div class="gh-dashboard5-chart-loading">
+                <div class="gh-dashboard-chart-loading">
                     <div class="gh-loading-spinner"></div>
                 </div>
             {{else}}
-                <div class="gh-dashboard5-chart-container">
-                    <div class="gh-dashboard5-chart-box">
+                <div class="gh-dashboard-chart-container">
+                    <div class="gh-dashboard-chart-box">
                         <EmberChart
                             @type={{this.chartType}}
                             @data={{this.chartData}}
@@ -26,8 +26,8 @@
                             @height={{110}} /> 
                     </div>
 
-                    <div id="gh-dashboard5-mix-tooltip" class="gh-dashboard5-tooltip">
-                        <div class="gh-dashboard5-tooltip-value">
+                    <div id="gh-dashboard-mix-tooltip" class="gh-dashboard-tooltip">
+                        <div class="gh-dashboard-tooltip-value">
                             -
                         </div>
                     </div> 
@@ -37,7 +37,7 @@
     </div>
 
     {{#if this.hasMultipleTiers }}
-        <div class="gh-dashboard5-select">
+        <div class="gh-dashboard-select">
             <PowerSelect
                 @selected={{this.selectedModeOption}}
                 @options={{this.modeOptions}}
