+++ app/components/dashboard/charts/recents.js
@@ -37,21 +37,6 @@ export default class Recents extends Component {
         return (this.selected === 'activity');
     }
 
-    get shouldDisplay() {
-        if (this.feature.improvedOnboarding) {
-            return true;
-        }
-
-        const isOwner = this.session.user?.isOwnerOnly;
-        const hasCompletedLaunchWizard = this.settings.get('editorIsLaunchComplete');
-
-        if (isOwner && !hasCompletedLaunchWizard) {
-            return false;
-        }
-
-        return true;
-    }
-
     get areMembersEnabled() {
         return this.dashboardStats.siteStatus?.membersEnabled;
     }
