--- app/components/dashboard/index.hbs
@@ -1,57 +1,0 @@
-<section class="gh-dashboard5-layout" {{did-insert this.onInsert}}>
-    {{#if this.isLoading }}
-        <GhLoadingSpinner />
-    {{else}}
-        {{#if this.areMembersEnabled}}
-            {{#if this.hasPaidTiers}}
-                <Dashboard::Charts::Overview />
-            {{/if}}
-
-            <div class="gh-dashboard5-group {{if this.isTotalMembersZero 'is-zero'}}">
-                <Dashboard::Charts::Anchor />
-
-                {{#if this.areNewslettersEnabled}}
-                    <Dashboard::Charts::Engagement />
-                {{/if}}
-
-                {{#if this.isTotalMembersZero}}
-                    <Dashboard::Parts::Zero />
-                {{/if}}
-            </div>
-        {{/if}}
-
-        <Dashboard::Charts::Recents />
-
-        <div class="gh-dashboard5-split gh-dashboard5-box is-secondary">
-            <Dashboard::Resources::Resources />
-            <Dashboard::Resources::Newsletter />
-        </div>
-        <div class="gh-dashboard5-split">
-            <Dashboard::Resources::StaffPicks />
-            <Dashboard::Resources::WhatsNew />
-            <Dashboard::Resources::Community />
-        </div>
-    {{/if}}
-
-    {{#unless this.isTotalMembersZero}}
-        <div class="gh-dashboard5-select">
-            <PowerSelect
-                @selected={{this.selectedDaysOption}}
-                @options={{this.daysOptions}}
-                @searchEnabled={{false}}
-                @onChange={{this.onDaysChange}}
-                @triggerComponent="gh-power-select/trigger"
-                @triggerClass="gh-contentfilter-menu-trigger"
-                @dropdownClass="gh-contentfilter-menu-dropdown is-narrow"
-                @matchTriggerWidth={{false}}
-                as |option|
-            >
-                {{#if option.name}}{{option.name}}{{else}}<span class="red">Unknown option</span>{{/if}}
-            </PowerSelect>
-        </div>
-    {{/unless}}
-</section>
-
-{{#if (enable-developer-experiments)}}
-    <Dashboard::Prototype::ControlPanel />
-{{/if}}
