+++ app/components/dashboard/parts/loading.hbs
@@ -1,3 +1,3 @@
-<div class="gh-dashboard5-zero">
+<div class="gh-dashboard-zero">
 
 </div>
