+++ app/components/dashboard/parts/metric.hbs
@@ -1,36 +1,36 @@
-<div class="gh-dashboard5-metric {{if @center "is-center"}} {{if @reverse "is-reverse"}} {{if @large "is-large"}}">
-    <div class="gh-dashboard5-metric-data">
+<div class="gh-dashboard-metric {{if @center "is-center"}} {{if @reverse "is-reverse"}} {{if @large "is-large"}}">
+    <div class="gh-dashboard-metric-data">
         {{#if @secondary}}
             {{#if @value}}
-                <div class="gh-dashboard5-metric-value {{if @secondary 'is-secondary'}}">
+                <div class="gh-dashboard-metric-value {{if @secondary 'is-secondary'}}">
                     <span class="value">{{@value}}</span>
                     {{#if @trends}}
                     <Dashboard::Parts::Percentage @percentage={{@percentage}}/>
                 {{/if}}
                 </div>
             {{/if}}
-            <h5 class="gh-dashboard5-metric-label {{if @secondary 'is-secondary'}}">
+            <h5 class="gh-dashboard-metric-label {{if @secondary 'is-secondary'}}">
                 {{@label}}
             </h5>
         {{else}}
             {{#if @reverse}}
                 {{#if @value}}
-                    <div class="gh-dashboard5-metric-value">
+                    <div class="gh-dashboard-metric-value">
                         <span class="value">{{@value}}</span>
                         {{#if @trends}}
                         <Dashboard::Parts::Percentage @percentage={{@percentage}}/>
                     {{/if}}
                     </div>
                 {{/if}}
-                <h5 class="gh-dashboard5-metric-label">
+                <h5 class="gh-dashboard-metric-label">
                     {{@label}}
                 </h5>
             {{else}}
-                <h5 class="gh-dashboard5-metric-label">
+                <h5 class="gh-dashboard-metric-label">
                     {{@label}}
                 </h5>
                 {{#if @value}}
-                    <div class="gh-dashboard5-metric-value">
+                    <div class="gh-dashboard-metric-value">
                         <span class="value">{{@value}}</span>
                         {{#if @trends}}
                             <Dashboard::Parts::Percentage @percentage={{@percentage}}/>
@@ -41,7 +41,7 @@
         {{/if}}
     </div>
     {{#if @extra}}
-        <div class="gh-dashboard5-metric-extra">
+        <div class="gh-dashboard-metric-extra">
             {{@extra}}
         </div>
     {{/if}}
