+++ app/components/dashboard/parts/percentage.hbs
@@ -1,7 +1,7 @@
 {{#if (gt @percentage 0) }}
-    <div class="gh-dashboard5-percentage is-positive">+{{ @percentage }}%</div>
+    <div class="gh-dashboard-percentage is-positive">+{{ @percentage }}%</div>
 {{else if (lt @percentage 0)}}
-    <div class="gh-dashboard5-percentage is-negative">{{ @percentage }}%</div>
+    <div class="gh-dashboard-percentage is-negative">{{ @percentage }}%</div>
 {{else}}
-    <div class="gh-dashboard5-percentage">0%</div>
+    <div class="gh-dashboard-percentage">0%</div>
 {{/if}}
