+++ app/components/dashboard/parts/zero.hbs
@@ -1,5 +1,5 @@
-<div class="gh-dashboard5-zero">
-    <div class="gh-dashboard5-zero-message">
+<div class="gh-dashboard-zero">
+    <div class="gh-dashboard-zero-message">
         <h4>Welcome to your Dashboard</h4>
         <p>You'll find member analytics here once<br />someone signs up.</p>
         <p><LinkTo @route="members">Add or import members &rarr;</LinkTo></p>
