+++ app/components/dashboard/resources/community.hbs
@@ -1,16 +1,16 @@
-<section class="gh-dashboard5-resource gh-dashboard5-community">
-    <a href="https://ghost.org/resources/community/" target="_blank" rel="noopener noreferrer" class="gh-dashboard5-resource-box">
-        <div class="gh-dashboard5-resource-title">
+<section class="gh-dashboard-resource gh-dashboard-community">
+    <a href="https://ghost.org/resources/community/" target="_blank" rel="noopener noreferrer" class="gh-dashboard-resource-box">
+        <div class="gh-dashboard-resource-title">
             <h4>Ghost Creator Community</h4>
         </div> 
-        <div class="gh-dashboard5-resource-body">
-            <div class="gh-dashboard5-list">
-                <div class="gh-dashboard5-list-body">
+        <div class="gh-dashboard-resource-body">
+            <div class="gh-dashboard-list">
+                <div class="gh-dashboard-list-body">
                     <p>Talk strategy.<br />Get advice.<br />Or just hang out.</p>
                 </div>
             </div>
         </div>
-        <div class="gh-dashboard5-resource-footer">
+        <div class="gh-dashboard-resource-footer">
             Share the journey &rarr;
         </div>
     </a>
