+++ app/components/dashboard/resources/newsletter.hbs
@@ -1,14 +1,14 @@
-<section class="gh-dashboard5-resource gh-dashboard5-newsletter" {{did-insert this.load}}>
-    <article class="gh-dashboard5-resource-box">
-        <div class="gh-dashboard5-resource-title">
+<section class="gh-dashboard-resource gh-dashboard-newsletter" {{did-insert this.load}}>
+    <article class="gh-dashboard-resource-box">
+        <div class="gh-dashboard-resource-title">
             <h4>Latest from the newsletter</h4>
         </div>
-        <div class="gh-dashboard5-resource-body">
+        <div class="gh-dashboard-resource-body">
             {{#if (not (or this.loading this.error))}}
-                <div class="gh-dashboard5-resource-bigarticle">
+                <div class="gh-dashboard-resource-bigarticle">
                     {{#each this.newsletters as |entry|}}
-                        <a class="gh-dashboard5-resource-smallarticle" href={{set-query-params entry.url utm_source='admin'}} target="_blank" rel="noopener noreferrer">
-                            <div class="gh-dashboard5-resource-text">
+                        <a class="gh-dashboard-resource-smallarticle" href={{set-query-params entry.url utm_source='admin'}} target="_blank" rel="noopener noreferrer">
+                            <div class="gh-dashboard-resource-text">
                                 <h3>{{entry.title}}</h3>
                                 <p>{{entry.excerpt}}</p>
                             </div>
@@ -17,8 +17,8 @@
                 </div>
             {{/if}}
         </div>
-        <div class="gh-dashboard5-resource-footer">
-            <a href="https://ghost.org/resources/newsletter/" target="_blank" class="gh-dashboard5-subscribe-button" rel="noopener noreferrer">Subscribe&nbsp;<span>to the newsletter&nbsp;</span>&rarr;</a>
+        <div class="gh-dashboard-resource-footer">
+            <a href="https://ghost.org/resources/newsletter/" target="_blank" class="gh-dashboard-subscribe-button" rel="noopener noreferrer">Subscribe&nbsp;<span>to the newsletter&nbsp;</span>&rarr;</a>
         </div>
     </article>
 </section>
