+++ app/components/dashboard/resources/resources.hbs
@@ -1,20 +1,20 @@
-<section class="gh-dashboard5-resource gh-dashboard5-resources" {{did-insert this.load}}>
-    <article class="gh-dashboard5-resource-box">
+<section class="gh-dashboard-resource gh-dashboard-resources" {{did-insert this.load}}>
+    <article class="gh-dashboard-resource-box">
         {{#if (not (or this.loading this.error))}}
-            <a href="{{this.resource.url}}" target="_blank" class="gh-dashboard5-resource-thumbnail" rel="noopener noreferrer" style={{html-safe (concat "background-image: url(" this.resource.feature_image ")")}} aria-label="Resource link"></a>        
-            <div class="gh-dashboard5-resource-contents">
-                <div class="gh-dashboard5-resource-title">
+            <a href="{{this.resource.url}}" target="_blank" class="gh-dashboard-resource-thumbnail" rel="noopener noreferrer" style={{html-safe (concat "background-image: url(" this.resource.feature_image ")")}} aria-label="Resource link"></a>        
+            <div class="gh-dashboard-resource-contents">
+                <div class="gh-dashboard-resource-title">
                     <h4>Resources</h4>
                 </div>
-                <div class="gh-dashboard5-resource-body">
-                    <a href="{{this.resource.url}}" target="_blank" class="gh-dashboard5-resource-bigarticle" rel="noopener noreferrer">
-                        <div class="gh-dashboard5-resource-text">
+                <div class="gh-dashboard-resource-body">
+                    <a href="{{this.resource.url}}" target="_blank" class="gh-dashboard-resource-bigarticle" rel="noopener noreferrer">
+                        <div class="gh-dashboard-resource-text">
                             <h3>{{this.resource.title}}</h3>
                             <p>{{this.resource.excerpt}}</p>
                         </div>
                     </a>
                 </div>
-                <div class="gh-dashboard5-resource-footer">
+                <div class="gh-dashboard-resource-footer">
                     <a href="https://ghost.org/resources/" target="_blank" rel="noopener noreferrer">Learn more &rarr;</a>
                 </div>
             </div>
