--- app/components/gh-members-chart.hbs
@@ -1,11 +1,0 @@
-<div class="gh-dashboard-chart-box {{if this.isSmall "small"}}" {{did-insert (perform this.fetchStatsTask)}}>    
-    {{#if this.stats}}
-        <EmberChart
-            @type={{this.type}}
-            @options={{this.chartOptions}}
-            @data={{this.chartData}}
-            @height={{300}} />
-    {{else}}
-        <GhLoadingSpinner />
-    {{/if}}
-</div>
