+++ app/templates/settings/labs.hbs
@@ -322,19 +322,6 @@
                         </div>
                     </div>
                 </div>
-                <div class="gh-expandable-block">
-                     <div class="gh-expandable-header">
-                         <div>
-                             <h4 class="gh-expandable-title">Dashboard 5.0</h4>
-                             <p class="gh-expandable-description">
-                                Increased amount of analytics with more useful / dynamic resources
-                             </p>
-                         </div>
-                         <div class="for-switch">
-                            <GhFeatureFlag @flag="dashboardV5" />
-                         </div>
-                     </div>
-                 </div>
                  <div class="gh-expandable-block">
                      <div class="gh-expandable-header">
                          <div>
