+++ tests/acceptance/dashboard-test.js
@@ -1,5 +1,5 @@
 import {authenticateSession, invalidateSession} from 'ember-simple-auth/test-support';
-import {currentURL, find, visit} from '@ember/test-helpers';
+import {currentURL, visit} from '@ember/test-helpers';
 import {describe, it} from 'mocha';
 import {enableLabsFlag} from '../helpers/labs-flag';
 import {expect} from 'chai';
@@ -30,28 +30,6 @@ describe('Acceptance: Dashboard', function () {
     it('/ redirects to /dashboard', async function () {
         await visit('/');
         expect(currentURL()).to.equal('/dashboard');
-    });
-
-    describe('members graphs', function () {
-        it('is shown when members exist', async function () {
-            this.server.createList('member', 5);
-            await visit('/dashboard');
-            expect(find('[data-test-dashboard-members-graphs]'), 'members graphs block').to.exist;
-        });
-
-        it('is hidden when no members exist', async function () {
-            this.server.db.members.remove();
-            await visit('/dashboard');
-            expect(find('[data-test-dashboard-members-graphs]'), 'members graphs block').to.not.exist;
-        });
-
-        it('is hidden when members is disabled', async function () {
-            this.server.createList('member', 5);
-            this.server.db.settings.update({key: 'members_signup_access'}, {value: 'none'});
-
-            await visit('/dashboard');
-            expect(find('[data-test-dashboard-members-graphs]'), 'members graphs block').to.not.exist;
-        });
     });
 
     describe('permissions', function () {
